feat: adicionando remoção de consultas
