+++ .env.example
@@ -1,0 +1,5 @@
+database=consultorio
+username=postgres
+password=12345
+host=localhost
+dialect=postgres
