+++ package.json
@@ -4,6 +4,7 @@
     "start": "npx tsx index.ts"
   },
   "dependencies": {
+    "dotenv": "^16.4.7",
     "pg": "^8.13.1",
     "pg-hstore": "^2.3.4",
     "prompt-sync": "^4.2.0",
