+++ setup/application.ts
@@ -1,17 +1,17 @@
 import db from "../database/database";
 import ConsultorioController from "../controllers/consultorio.controller";
-import NewConsultorioView from "../views/new.consultorio.view";
+import ConsultorioView from "../views/consultorio.view";
 import PromptSync from "prompt-sync";
 
 class Application {
   private prompt: any;
   private consultorioController: ConsultorioController;
-  private consultorioView: NewConsultorioView;
+  private consultorioView: ConsultorioView;
 
   constructor() {
     this.prompt = PromptSync();
     this.consultorioController = new ConsultorioController();
-    this.consultorioView = new NewConsultorioView(
+    this.consultorioView = new ConsultorioView(
       this.prompt,
       this.consultorioController,
     );
